services/kaapana-core/kube-helm/docker/files/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
python-dotenv==0.13.0
elasticsearch
PyYAML==5.3.1